+++ dom/media/GraphDriver.h
@@ -79,7 +79,7 @@ namespace dom {
 enum class AudioContextOperation;
 }
 
-struct GraphInterface {
+struct GraphInterface : public nsISupports {
   /**
    * Object returned from OneIteration() instructing the iterating GraphDriver
    * what to do.
@@ -173,8 +173,6 @@ struct GraphInterface {
       mResult.as<SwitchDriver>().Switched();
     }
   };
-
-  NS_INLINE_DECL_PURE_VIRTUAL_REFCOUNTING
 
   /* Called on the graph thread when there is new output data for listeners.
    * This is the mixed audio output of this MediaTrackGraph. */
@@ -427,7 +425,7 @@ class ThreadedDriver : public GraphDriver {
    * Runs main control loop on the graph thread. Normally a single invocation
    * of this runs for the entire lifetime of the graph thread.
    */
-  void RunThread();
+  virtual void RunThread();
   friend class MediaTrackGraphInitThreadRunnable;
   uint32_t IterationDuration() override { return MEDIA_GRAPH_TARGET_PERIOD_MS; }
 
@@ -498,6 +496,8 @@ class OfflineClockDriver : public ThreadedDriver {
                      GraphTime aSlice);
   virtual ~OfflineClockDriver();
   OfflineClockDriver* AsOfflineClockDriver() override { return this; }
+
+  void RunThread() override;
 
  protected:
   TimeDuration WaitInterval() override { return TimeDuration(); }