+++ mobile/android/components/extensions/ext-tabs.js
@@ -309,58 +309,64 @@ this.tabs = class extends ExtensionAPI {
           },
         }).api(),
 
-        async create(createProperties) {
-          let principal = context.principal;
-          let window =
-            createProperties.windowId !== null
-              ? windowTracker.getWindow(createProperties.windowId, context)
-              : windowTracker.topWindow;
+        async create({
+          active,
+          discarded,
+          index,
+          openInReaderMode,
+          pinned,
+          title,
+          url,
+        } = {}) {
+          if (active === null) {
+            active = true;
+          }
 
-          let { BrowserApp } = window;
-          let url;
+          tabListener.initTabReady();
 
-          if (createProperties.url !== null) {
-            url = context.uri.resolve(createProperties.url);
+          if (url !== null) {
+            url = context.uri.resolve(url);
 
             if (!context.checkLoadURL(url, { dontReportErrors: true })) {
               return Promise.reject({ message: `Illegal URL: ${url}` });
             }
-          } else {
-            // Falling back to system here as about:newtab requires it, however is safe.
-            principal = Services.scriptSecurityManager.getSystemPrincipal();
           }
 
-          let options = {};
+          const nativeTab = await GeckoViewTabBridge.createNewTab({
+            extensionId: context.extension.id,
+            createProperties: {
+              active,
+              discarded,
+              index,
+              openInReaderMode,
+              pinned,
+              url,
+            },
+          });
 
-          let active = true;
-          if (createProperties.active !== null) {
-            active = createProperties.active;
-          }
-          options.selected = active;
+          const { browser } = nativeTab;
 
-          if (createProperties.index !== null) {
-            options.tabIndex = createProperties.index;
-          }
+          let flags = Ci.nsIWebNavigation.LOAD_FLAGS_NONE;
 
           // Make sure things like about:blank URIs never inherit,
           // and instead always get a NullPrincipal.
-          if (url && url.startsWith("about:")) {
-            options.disallowInheritPrincipal = true;
+          if (url !== null) {
+            tabListener.initializingTabs.add(nativeTab);
           } else {
-            options.triggeringPrincipal = context.principal;
+            url = "about:blank";
+            flags |= Ci.nsIWebNavigation.LOAD_FLAGS_DISALLOW_INHERIT_PRINCIPAL;
           }
 
-          options.parentId = BrowserApp.selectedTab.id;
+          browser.loadURI(url, {
+            flags,
+            // GeckoView doesn't support about:newtab so we don't need to worry
+            // about using the system principal here.
+            triggeringPrincipal: context.principal,
+          });
 
-          tabListener.initTabReady();
-          options.triggeringPrincipal = principal;
-
-          options.extensionId = context.extension.id;
-          options.url = url;
-
-          let nativeTab = await GeckoViewTabBridge.createNewTab(options);
-          if (createProperties.url) {
-            tabListener.initializingTabs.add(nativeTab);
+          if (active) {
+            const newWindow = browser.ownerGlobal;
+            mobileWindowTracker.setTabActive(newWindow, true);
           }
 
           return tabManager.convert(nativeTab);
@@ -386,28 +392,44 @@ this.tabs = class extends ExtensionAPI {
           );
         },
 
-        async update(tabId, updateProperties) {
-          let nativeTab = getTabOrActive(tabId);
+        async update(
+          tabId,
+          { active, autoDiscardable, highlighted, muted, pinned, url } = {}
+        ) {
+          const nativeTab = getTabOrActive(tabId);
+          const window = nativeTab.browser.ownerGlobal;
 
-          let { BrowserApp } = nativeTab.browser.ownerGlobal;
-
-          if (updateProperties.url !== null) {
-            let url = context.uri.resolve(updateProperties.url);
+          if (url !== null) {
+            url = context.uri.resolve(url);
 
             if (!context.checkLoadURL(url, { dontReportErrors: true })) {
               return Promise.reject({ message: `Illegal URL: ${url}` });
             }
+          }
 
-            let options = {
+          await GeckoViewTabBridge.updateTab({
+            window,
+            extensionId: context.extension.id,
+            updateProperties: {
+              active,
+              autoDiscardable,
+              highlighted,
+              muted,
+              pinned,
+              url,
+            },
+          });
+
+          if (url !== null) {
+            nativeTab.browser.loadURI(url, {
               triggeringPrincipal: context.principal,
-            };
-            nativeTab.browser.loadURI(url, options);
+            });
           }
 
-          if (updateProperties.active) {
-            BrowserApp.selectTab(nativeTab);
+          // FIXME: openerTabId, successorTabId
+          if (active) {
+            mobileWindowTracker.setTabActive(window, true);
           }
-          // FIXME: highlighted/selected, muted, pinned, openerTabId, successorTabId
 
           return tabManager.convert(nativeTab);
         },