+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/WebExtensionController.java
@@ -29,7 +29,7 @@ public class WebExtensionController {
 
     private DebuggerDelegate mDebuggerDelegate;
     private PromptDelegate mPromptDelegate;
-    private final WebExtension.Listener mListener;
+    private final WebExtension.Listener<WebExtension.TabDelegate> mListener;
 
     // Map [ portId -> Message ]
     private final MultiMap<Long, Message> mPendingPortMessages;
@@ -181,6 +181,16 @@ public class WebExtensionController {
         public WebExtension.ActionDelegate getActionDelegate() {
             return mListener.getActionDelegate(mExtension);
         }
+
+        @Override
+        public void onTabDelegate(final WebExtension.TabDelegate delegate) {
+            mListener.setTabDelegate(mExtension, delegate);
+        }
+
+        @Override
+        public WebExtension.TabDelegate getTabDelegate() {
+            return mListener.getTabDelegate(mExtension);
+        }
     }
 
     // TODO: remove once we remove GeckoRuntime#registerWebExtension
@@ -188,6 +198,10 @@ public class WebExtensionController {
         return new DelegateController(extension);
     }
 
+    /**
+     * @deprecated Use {@link WebExtension.TabDelegate} and {@link WebExtension.SessionTabDelegate}.
+     */
+    @Deprecated
     public interface TabDelegate {
         /**
          * Called when tabs.create is invoked, this method returns a *newly-created* session
@@ -225,11 +239,24 @@ public class WebExtensionController {
         }
     }
 
+    /**
+     * @deprecated Use {@link WebExtension#getTabDelegate} and
+     * {@link WebExtension.SessionController#getTabDelegate}.
+     * @return The {@link TabDelegate} instance.
+     */
     @UiThread
-    public @Nullable TabDelegate getTabDelegate() {
+    @Nullable
+    @Deprecated
+    public TabDelegate getTabDelegate() {
         return mListener.getTabDelegate();
     }
 
+    /**
+     * @deprecated Use {@link WebExtension#setTabDelegate} and
+     * {@link WebExtension.SessionController#setTabDelegate}.
+     * @param delegate {@link TabDelegate} instance.
+     */
+    @Deprecated
     @UiThread
     public void setTabDelegate(final @Nullable TabDelegate delegate) {
         mListener.setTabDelegate(delegate);
@@ -497,15 +524,16 @@ public class WebExtensionController {
             }
         };
 
-        unregisterWebExtension(extension);
-
         final GeckoBundle bundle = new GeckoBundle(1);
         bundle.putString("webExtensionId", extension.id);
 
         EventDispatcher.getInstance().dispatch("GeckoView:WebExtension:Uninstall",
                 bundle, result);
 
-        return result;
+        return result.then(success -> {
+            unregisterWebExtension(extension);
+            return GeckoResult.fromValue(success);
+        });
     }
 
     @Retention(RetentionPolicy.SOURCE)
@@ -665,7 +693,7 @@ public class WebExtensionController {
     }
 
     /* package */ WebExtensionController(final GeckoRuntime runtime) {
-        mListener = new WebExtension.Listener(runtime);
+        mListener = new WebExtension.Listener<>(runtime);
         mPendingPortMessages = new MultiMap<>();
         mPendingMessages = new MultiMap<>();
         mExtensions.setObserver(mInternals);
@@ -685,6 +713,15 @@ public class WebExtensionController {
             return;
         } else if ("GeckoView:WebExtension:PortMessage".equals(event)) {
             portMessage(message);
+            return;
+        } else if ("GeckoView:WebExtension:NewTab".equals(event)) {
+            newTab(bundle, callback);
+            return;
+        } else if ("GeckoView:WebExtension:UpdateTab".equals(event)) {
+            updateTab(bundle, callback, session);
+            return;
+        } else if ("GeckoView:WebExtension:CloseTab".equals(event)) {
+            closeTab(bundle, callback, session);
             return;
         } else if ("GeckoView:BrowserAction:Update".equals(event)) {
             actionUpdate(bundle, session, WebExtension.Action.TYPE_BROWSER_ACTION);
@@ -721,9 +758,8 @@ public class WebExtensionController {
         }
 
         final GeckoBundle senderBundle = bundle.getBundle("sender");
-        final String extensionId = senderBundle.getString("extensionId");
 
-        mExtensions.get(extensionId).accept(extension -> {
+        extensionFromBundle(senderBundle).accept(extension -> {
             final WebExtension.MessageSender sender = fromBundle(extension, senderBundle, session);
             if (sender == null) {
                 if (callback != null) {
@@ -818,11 +854,24 @@ public class WebExtensionController {
         });
     }
 
+    /* package */ void newTab(final GeckoBundle message, final EventCallback callback) {
+        newTab(message, callback, null);
+    }
+
+    // TODO: remove legacyDelegate Bug 1618987
     /* package */ void newTab(final GeckoBundle message, final EventCallback callback,
-                              final TabDelegate delegate) {
-        mExtensions.get(message.getString("extensionId")).then(extension ->
-            delegate.onNewTab(extension, message.getString("uri"))
-        ).accept(session -> {
+                              final TabDelegate legacyDelegate) {
+        extensionFromBundle(message).then(extension -> {
+            final WebExtension.TabDelegate delegate = mListener.getTabDelegate(extension);
+            final WebExtension.CreateTabDetails details =
+                    new WebExtension.CreateTabDetails(message.getBundle("createProperties"));
+            if (delegate != null) {
+                return delegate.onNewTab(extension, details);
+            } else if (legacyDelegate != null) {
+                return legacyDelegate.onNewTab(extension, details.url);
+            }
+            return null;
+        }).accept(session -> {
             if (session == null) {
                 callback.sendSuccess(null);
                 return;
@@ -838,17 +887,47 @@ public class WebExtensionController {
         });
     }
 
+    /* package */ void updateTab(final GeckoBundle message,
+                                 final EventCallback callback,
+                                 final GeckoSession session) {
+        extensionFromBundle(message).then(extension -> {
+            final WebExtension.SessionTabDelegate delegate = session.getWebExtensionController()
+                    .getTabDelegate(extension);
+            if (delegate == null) {
+                return GeckoResult.fromValue(AllowOrDeny.DENY);
+            }
+            return delegate.onUpdateTab(extension, session,
+                    new WebExtension.UpdateTabDetails(message.getBundle("updateProperties")));
+        }).accept(value -> {
+            if (value == AllowOrDeny.ALLOW) {
+                callback.sendSuccess(null);
+            } else {
+                callback.sendError(null);
+            }
+        });
+    }
+
     /* package */ void closeTab(final GeckoBundle message,
                                 final EventCallback callback,
-                                final TabDelegate delegate,
                                 final GeckoSession session) {
-        mExtensions.get(message.getString("extensionId")).then(
-            extension -> delegate.onCloseTab(extension, session),
-            // On uninstall, we close all extension pages, in that case
-            // the extension object may be gone already so we can't
-            // send it to the delegate
-            exception -> delegate.onCloseTab(null, session)
-        ).accept(value -> {
+        closeTab(message, callback, session, null);
+    }
+
+    // TODO: remove legacyDelegate Bug 1618987
+    /* package */ void closeTab(final GeckoBundle message,
+                                final EventCallback callback,
+                                final GeckoSession session,
+                                final TabDelegate legacyDelegate) {
+        extensionFromBundle(message).then(extension -> {
+            final WebExtension.SessionTabDelegate delegate = session.getWebExtensionController()
+                    .getTabDelegate(extension);
+            if (delegate != null) {
+                return delegate.onCloseTab(extension, session);
+            } else if (legacyDelegate != null) {
+                return legacyDelegate.onCloseTab(extension, session);
+            }
+            return GeckoResult.fromValue(AllowOrDeny.DENY);
+        }).accept(value -> {
             if (value == AllowOrDeny.ALLOW) {
                 callback.sendSuccess(null);
             } else {
@@ -875,6 +954,7 @@ public class WebExtensionController {
     /* package */ void unregisterWebExtension(final WebExtension webExtension) {
         mExtensions.remove(webExtension.id);
         webExtension.setDelegateController(null);
+        mListener.unregisterWebExtension(webExtension);
 
         // Some ports may still be open so we need to go through the list and close all of the
         // ports tied to this web extension
@@ -972,7 +1052,8 @@ public class WebExtensionController {
         WebExtension.MessageDelegate delegate = null;
 
         if (sender.session != null) {
-            delegate = sender.session.getMessageDelegate(sender.webExtension, nativeApp);
+            delegate = sender.session.getWebExtensionController()
+                    .getMessageDelegate(sender.webExtension, nativeApp);
         } else if (sender.environmentType == WebExtension.MessageSender.ENV_TYPE_EXTENSION) {
             delegate = mListener.getMessageDelegate(sender.webExtension, nativeApp);
         }
@@ -1101,7 +1182,7 @@ public class WebExtensionController {
             return mListener.getActionDelegate(extension);
         }
 
-        return session.getWebExtensionActionDelegate(extension);
+        return session.getWebExtensionController().getActionDelegate(extension);
     }
 
     private void actionUpdate(final GeckoBundle message, final GeckoSession session,